requirements.txt
spacy>=3.7.0
gliner>=0.2.0
transformers>=4.45.0
torch==2.6.0+cu118
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
rapidfuzz>=3.6.0
pdfplumber>=0.11.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
tqdm>=4.67.0
gradio>=4.0.0

# LELA core (optional)
bm25s[full]>=0.2.13
PyStemmer>=2.0.0

# LELA vLLM (optional, requires GPU)
vllm>=0.9.0
openai>=1.84.0

